+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 ## v1.1.0.5 (XX-XX-2015)
  * Changed channel updater to use the search API instead of the channel's playlist API (This allows more up-to-date updating)
  * Fixed bug where channels would stop updating
+ * Added locales
 
 ## v1.1.0.4 (12-26-2015)
  * Decreased update interval from every 10 minutes to every 5 minutes