+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
  * Fixed bug where channels would stop updating
  * Added locales/languages
  * Added a 'load more' button in the information popup
+ * Added a 'view subscriptions' button in the edit settings tab
 
 ## v1.1.0.4 (12-26-2015)
  * Decreased update interval from every 10 minutes to every 5 minutes