+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Changelog
 =====
-## v1.1.0.5 (XX-XX-2015)
+## v1.1.0.5 (02-02-2015)
  * Changed channel updater to use the search API instead of the channel's playlist API (This allows more up-to-date updating)
  * Fixed bug where channels would stop updating
  * Added locales/languages